refactor: 상수 분리
- 야채 정보를 상수로 분리
