+++ client/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './getAssetPath';
 export * from './makeChart';
+export * from './constants';
